+++ src/plugins/mb_expander.cpp
@@ -900,8 +900,8 @@ namespace lsp
                     }
 
                     // Update transfer function for equalizer
-                    b->sEQ[0].freq_chart(0, b->vTr, vFreqs, mb_expander_base_metadata::FFT_MESH_POINTS);
-                    b->sEQ[0].freq_chart(1, vTr, vFreqs, mb_expander_base_metadata::FFT_MESH_POINTS);
+                    b->sEQ[0].freq_chart(size_t(0), b->vTr, vFreqs, mb_expander_base_metadata::FFT_MESH_POINTS);
+                    b->sEQ[0].freq_chart(size_t(1), vTr, vFreqs, mb_expander_base_metadata::FFT_MESH_POINTS);
                     dsp::pcomplex_mul2(b->vTr, vTr, mb_expander_base_metadata::FFT_MESH_POINTS);
                     dsp::pcomplex_mod(b->vTr, b->vTr, mb_expander_base_metadata::FFT_MESH_POINTS);
